+++ src/plugin.hpp
@@ -29,6 +29,8 @@ extern Model* modelMixOffset;
 extern Model* modelMixPan;
 extern Model* modelMixSend;
 extern Model* modelMixSolo;
+extern Model* modelMultiMerge;
+extern Model* modelMultiSplit;
 extern Model* modelNORS_IQ;
 extern Model* modelNORSIQChord2Scale;
 extern Model* modelPolyClone;